Converted RandGraphData_Generator to ts & return new type "GeneratedGraphData"
